# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ControlCreate": {
        "additionalProperties": false,
        "properties": {
          "applicability_criteria": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Applicability Criteria",
            "type": "string"
          },
          "change_reason": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Change Reason",
            "type": "string"
          },
          "description": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "engineering_owner": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Engineering Owner",
            "type": "string"
          },
          "exception_eligible": {
            "title": "Exception Eligible",
            "type": "boolean"
          },
          "framework_refs": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Framework Refs",
            "type": "array"
          },
          "id": {
            "pattern": "^CTL-[A-Z0-9-]{3,60}$",
            "title": "Id",
            "type": "string"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Limitations",
            "type": "array"
          },
          "name": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "operational_owner": {
            "default": "Cloud Engineering",
            "maxLength": 256,
            "minLength": 1,
            "title": "Operational Owner",
            "type": "string"
          },
          "origin": {
            "default": "PROPOSED",
            "enum": [
              "PROPOSED",
              "IMPORTED",
              "EXTERNALLY_MANAGED"
            ],
            "title": "Origin",
            "type": "string"
          },
          "prevention_boundary": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Prevention Boundary",
            "type": "string"
          },
          "providers": {
            "items": {
              "enum": [
                "azure",
                "aws"
              ],
              "type": "string"
            },
            "maxItems": 2,
            "minItems": 1,
            "title": "Providers",
            "type": "array"
          },
          "rationale": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          },
          "resource_types": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Resource Types",
            "type": "array"
          },
          "security_objective": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Security Objective",
            "type": "string"
          },
          "security_owner": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Security Owner",
            "type": "string"
          },
          "severity": {
            "enum": [
              "CRITICAL",
              "HIGH",
              "MEDIUM",
              "LOW"
            ],
            "title": "Severity",
            "type": "string"
          },
          "source_evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "maxItems": 50,
            "title": "Source Evidence",
            "type": "array"
          }
        },
        "required": [
          "name",
          "description",
          "security_objective",
          "rationale",
          "severity",
          "providers",
          "resource_types",
          "applicability_criteria",
          "security_owner",
          "engineering_owner",
          "exception_eligible",
          "prevention_boundary",
          "id",
          "change_reason"
        ],
        "title": "ControlCreate",
        "type": "object"
      },
      "ControlRevisionUpdate": {
        "additionalProperties": false,
        "properties": {
          "applicability_criteria": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Applicability Criteria",
            "type": "string"
          },
          "change_reason": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Change Reason",
            "type": "string"
          },
          "description": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "engineering_owner": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Engineering Owner",
            "type": "string"
          },
          "exception_eligible": {
            "title": "Exception Eligible",
            "type": "boolean"
          },
          "expected_lock_version": {
            "title": "Expected Lock Version",
            "type": "integer"
          },
          "framework_refs": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Framework Refs",
            "type": "array"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Limitations",
            "type": "array"
          },
          "name": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "prevention_boundary": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Prevention Boundary",
            "type": "string"
          },
          "providers": {
            "items": {
              "enum": [
                "azure",
                "aws"
              ],
              "type": "string"
            },
            "maxItems": 2,
            "minItems": 1,
            "title": "Providers",
            "type": "array"
          },
          "rationale": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          },
          "resource_types": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Resource Types",
            "type": "array"
          },
          "security_objective": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Security Objective",
            "type": "string"
          },
          "security_owner": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Security Owner",
            "type": "string"
          },
          "severity": {
            "enum": [
              "CRITICAL",
              "HIGH",
              "MEDIUM",
              "LOW"
            ],
            "title": "Severity",
            "type": "string"
          },
          "source_evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "maxItems": 50,
            "title": "Source Evidence",
            "type": "array"
          }
        },
        "required": [
          "name",
          "description",
          "security_objective",
          "rationale",
          "severity",
          "providers",
          "resource_types",
          "applicability_criteria",
          "security_owner",
          "engineering_owner",
          "exception_eligible",
          "prevention_boundary",
          "change_reason",
          "expected_lock_version"
        ],
        "title": "ControlRevisionUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LockedCommand": {
        "additionalProperties": false,
        "properties": {
          "expected_lock_version": {
            "title": "Expected Lock Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_lock_version"
        ],
        "title": "LockedCommand",
        "type": "object"
      },
...
      "NewRevision": {
        "additionalProperties": false,
        "properties": {
          "change_reason": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Change Reason",
            "type": "string"
          }
        },
        "required": [
          "change_reason"
        ],
        "title": "NewRevision",
        "type": "object"
      },
...
      "ReasonCommand": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ReasonCommand",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Security intent -> governed, verifiable preventive controls. Fixture data only; no cloud mutations. Delivery is a local GitOps handoff bundle and a clearly labelled mock pipeline.",
    "title": "Cloud Security Control Engineering Platform (local MVP)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/control-revisions/{revision_id}": {
      "delete": {
        "operationId": "delete_control_revision_api_v1_control_revisions__revision_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Control Revision",
        "tags": [
          "controls"
        ]
      },
      "put": {
        "operationId": "update_control_revision_api_v1_control_revisions__revision_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlRevisionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Control Revision",
        "tags": [
          "controls"
        ]
      }
    },
    "/api/v1/control-revisions/{revision_id}/submit": {
      "post": {
        "operationId": "submit_control_revision_api_v1_control_revisions__revision_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LockedCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Control Revision",
        "tags": [
          "controls"
        ]
      }
    },
    "/api/v1/controls": {
      "get": {
        "operationId": "list_controls_api_v1_controls_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Controls",
        "tags": [
          "controls"
        ]
      },
      "post": {
        "operationId": "create_control_api_v1_controls_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Control",
        "tags": [
          "controls"
        ]
      }
    },
    "/api/v1/controls/{control_id}": {
      "get": {
        "operationId": "get_control_api_v1_controls__control_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "control_id",
            "required": true,
            "schema": {
              "title": "Control Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Control",
        "tags": [
          "controls"
        ]
      }
    },
    "/api/v1/controls/{control_id}/implementations": {
      "post": {
        "operationId": "create_implementation_api_v1_controls__control_id__implementations_post",
        "parameters": [
          {
            "in": "path",
            "name": "control_id",
            "required": true,
            "schema": {
              "title": "Control Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImplementationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Implementation",
        "tags": [
          "controls"
        ]
      }
    },
    "/api/v1/controls/{control_id}/retire": {
      "post": {
        "operationId": "retire_control_api_v1_controls__control_id__retire_post",
        "parameters": [
          {
            "in": "path",
            "name": "control_id",
            "required": true,
            "schema": {
              "title": "Control Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retire Control",
        "tags": [
          "controls"
        ]
      }
    },
    "/api/v1/controls/{control_id}/revisions": {
      "post": {
        "operationId": "new_control_revision_api_v1_controls__control_id__revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "control_id",
            "required": true,
            "schema": {
              "title": "Control Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewRevision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "New Control Revision",
        "tags": [
          "controls"
        ]
      }
    },
...
    "/api/v1/implementation-revisions/{rev_id}": {
      "delete": {
        "operationId": "delete_implementation_revision_api_v1_implementation_revisions__rev_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "rev_id",
            "required": true,
            "schema": {
              "title": "Rev Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Implementation Revision",
        "tags": [
          "controls"
        ]
      },
      "get": {
        "operationId": "get_implementation_revision_api_v1_implementation_revisions__rev_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "rev_id",
            "required": true,
            "schema": {
              "title": "Rev Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Implementation Revision",
        "tags": [
          "controls"
        ]
      },
      "put": {
        "operationId": "update_implementation_revision_api_v1_implementation_revisions__rev_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "rev_id",
            "required": true,
            "schema": {
              "title": "Rev Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImplementationRevisionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Implementation Revision",
        "tags": [
          "controls"
        ]
      }
    },
    "/api/v1/implementation-revisions/{rev_id}/submit": {
      "post": {
        "operationId": "submit_implementation_revision_api_v1_implementation_revisions__rev_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "rev_id",
            "required": true,
            "schema": {
              "title": "Rev Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LockedCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Implementation Revision",
        "tags": [
          "controls"
        ]
      }
    },
    "/api/v1/implementation-revisions/{rev_id}/validate": {
      "post": {
        "operationId": "validate_implementation_revision_api_v1_implementation_revisions__rev_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "rev_id",
            "required": true,
            "schema": {
              "title": "Rev Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Implementation Revision",
        "tags": [
          "controls"
        ]
      }
    },
    "/api/v1/implementations/{impl_id}": {
      "get": {
        "operationId": "get_implementation_api_v1_implementations__impl_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "impl_id",
            "required": true,
            "schema": {
              "title": "Impl Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Implementation",
        "tags": [
          "controls"
        ]
      }
    },
    "/api/v1/implementations/{impl_id}/revisions": {
      "post": {
        "operationId": "new_implementation_revision_api_v1_implementations__impl_id__revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "impl_id",
            "required": true,
            "schema": {
              "title": "Impl Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImplementationRevisionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "New Implementation Revision",
        "tags": [
          "controls"
        ]
      }
    },
...
  }
}